refactor: rename module to avoid hiding base types
