+++ src/tryceratops/runners.py
@@ -5,9 +5,9 @@ from typing import List, Set, Type
 
 from tryceratops.analyzers import BaseAnalyzer, get_analyzer_chain
 from tryceratops.fixers import BaseFixer, get_fixers_chain
+from tryceratops.parsing_types import ParsedFilesType
 from tryceratops.processors import Processor
 from tryceratops.settings import FileFilter, GlobalSettings
-from tryceratops.types import ParsedFilesType
 from tryceratops.violations import Violation
 
 logger = logging.getLogger(__name__)
